move punkt download to vector store where it's actually used
