--- sage/.sample-env
@@ -1,3 +1,0 @@
-OPENAI_API_KEY=
-PINECONE_API_KEY=
-GITHUB_TOKEN=
